fix(condo): DOMA-7148 added createUserAndSendLoginData function
